requirements.txt
streamlit
pandas
requests
beautifulsoup4
openai
python-dotenv
supabase
